Generalize CLF generation interface
