+++ opencolorio_config_aces/clf/transforms/__init__.py
@@ -1,6 +1,16 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright Contributors to the OpenColorIO Project.
 
-from .utility import generate_clf
+from .utilities import (
+    create_matrix,
+    create_conversion_matrix,
+    create_gamma,
+    generate_clf
+)
 
-__all__ = ["generate_clf"]
+__all__ = [
+    "create_matrix",
+    "create_conversion_matrix",
+    "create_gamma",
+    "generate_clf",
+]
